--- Input BEV/bev_data.csv
@@ -1,193 +1,0 @@
-;Etron;Audi
-1;0;0
-2;0;0
-3;0;0
-4;0;0
-5;0;0
-6;0;0
-7;0;0
-8;0;0
-9;0;0
-10;0;0
-11;0;0
-12;0;0
-13;0;0
-14;0;0
-15;0;0
-16;0;0
-17;0;0
-18;0;0
-19;0;0
-20;0;0
-21;0;0
-22;0;0
-23;0;0
-24;0;0
-25;0;0
-26;0;0
-27;0;0
-28;0;0
-29;1;0
-30;1;0
-31;1;0
-32;1;0
-33;1;0
-34;1;0
-35;1;1
-36;1;1
-37;1;1
-38;1;1
-39;1;1
-40;1;1
-41;1;1
-42;1;1
-43;1;1
-44;1;1
-45;1;1
-46;1;1
-47;1;1
-48;1;1
-49;1;1
-50;1;0
-51;1;0
-52;1;0
-53;1;0
-54;1;0
-55;1;0
-56;1;0
-57;1;0
-58;1;0
-59;1;1
-60;1;1
-61;1;1
-62;1;1
-63;1;1
-64;1;1
-65;1;1
-66;1;1
-67;1;1
-68;1;1
-69;1;1
-70;1;1
-71;1;1
-72;0;0
-73;0;0
-74;0;0
-75;0;0
-76;0;0
-77;0;0
-78;0;0
-79;0;0
-80;0;0
-81;0;0
-82;0;0
-83;0;0
-84;0;0
-85;0;0
-86;0;0
-87;0;0
-88;0;0
-89;0;0
-90;0;0
-91;0;0
-92;0;0
-93;0;0
-94;0;0
-95;0;0
-96;0;0
-97;0;0
-98;0;0
-99;0;0
-100;0;0
-101;0;0
-102;0;0
-103;0;0
-104;0;0
-105;0;0
-106;0;0
-107;0;0
-108;0;0
-109;0;0
-110;0;0
-111;0;0
-112;0;0
-113;0;0
-114;0;0
-115;0;0
-116;0;0
-117;0;0
-118;0;0
-119;0;0
-120;0;0
-121;0;0
-122;0;0
-123;0;0
-124;0;0
-125;1;0
-126;1;0
-127;1;0
-128;1;0
-129;1;0
-130;1;0
-131;1;1
-132;1;1
-133;1;1
-134;1;1
-135;1;1
-136;1;1
-137;1;1
-138;1;1
-139;1;1
-140;1;1
-141;1;1
-142;1;1
-143;1;1
-144;1;1
-145;1;1
-146;1;0
-147;1;0
-148;1;0
-149;1;0
-150;1;0
-151;1;0
-152;1;0
-153;1;0
-154;1;0
-155;1;1
-156;1;1
-157;1;1
-158;1;1
-159;1;1
-160;1;1
-161;1;1
-162;1;1
-163;1;1
-164;1;1
-165;1;1
-166;1;1
-167;1;1
-168;0;0
-169;0;0
-170;0;0
-171;0;0
-172;0;0
-173;0;0
-174;0;0
-175;0;0
-176;0;0
-177;0;0
-178;0;0
-179;0;0
-180;0;0
-181;0;0
-182;0;0
-183;0;0
-184;0;0
-185;0;0
-186;0;0
-187;0;0
-188;0;0
-;0;0
-;0;0
-;0;0
-;0;0
