+++ Input Variable Load/valo_data.csv
@@ -1,0 +1,193 @@
+;Etron;Audi;K2i
+1;0;0;1
+2;0;0;1
+3;0;0;1
+4;0;0;1
+5;0;0;1
+6;0;0;1
+7;0;0;1
+8;0;0;1
+9;0;0;1
+10;0;0;1
+11;0;0;1
+12;0;0;1
+13;0;0;1
+14;0;0;1
+15;0;0;1
+16;0;0;1
+17;0;0;1
+18;0;0;1
+19;0;0;1
+20;0;0;1
+21;0;0;1
+22;0;0;1
+23;0;0;1
+24;0;0;1
+25;0;0;1
+26;0;0;1
+27;0;0;1
+28;0;0;1
+29;1;0;1
+30;1;0;1
+31;1;0;1
+32;1;0;1
+33;1;0;1
+34;1;0;1
+35;1;1;1
+36;1;1;1
+37;1;1;1
+38;1;1;1
+39;1;1;1
+40;1;1;1
+41;1;1;1
+42;1;1;1
+43;1;1;1
+44;1;1;1
+45;1;1;1
+46;1;1;1
+47;1;1;1
+48;1;1;1
+49;1;1;1
+50;1;0;1
+51;1;0;1
+52;1;0;1
+53;1;0;1
+54;1;0;1
+55;1;0;1
+56;1;0;1
+57;1;0;1
+58;1;0;1
+59;1;1;1
+60;1;1;1
+61;1;1;1
+62;1;1;1
+63;1;1;1
+64;1;1;1
+65;1;1;1
+66;1;1;1
+67;1;1;1
+68;1;1;1
+69;1;1;1
+70;1;1;1
+71;1;1;1
+72;0;0;1
+73;0;0;1
+74;0;0;1
+75;0;0;1
+76;0;0;1
+77;0;0;1
+78;0;0;1
+79;0;0;1
+80;0;0;1
+81;0;0;1
+82;0;0;1
+83;0;0;1
+84;0;0;1
+85;0;0;1
+86;0;0;1
+87;0;0;1
+88;0;0;1
+89;0;0;1
+90;0;0;1
+91;0;0;1
+92;0;0;1
+93;0;0;1
+94;0;0;1
+95;0;0;1
+96;0;0;1
+97;0;0;1
+98;0;0;1
+99;0;0;1
+100;0;0;1
+101;0;0;1
+102;0;0;1
+103;0;0;1
+104;0;0;1
+105;0;0;1
+106;0;0;1
+107;0;0;1
+108;0;0;1
+109;0;0;1
+110;0;0;1
+111;0;0;1
+112;0;0;1
+113;0;0;1
+114;0;0;1
+115;0;0;1
+116;0;0;1
+117;0;0;1
+118;0;0;1
+119;0;0;1
+120;0;0;1
+121;0;0;1
+122;0;0;1
+123;0;0;1
+124;0;0;1
+125;1;0;1
+126;1;0;1
+127;1;0;1
+128;1;0;1
+129;1;0;1
+130;1;0;1
+131;1;1;1
+132;1;1;1
+133;1;1;1
+134;1;1;1
+135;1;1;1
+136;1;1;1
+137;1;1;1
+138;1;1;1
+139;1;1;1
+140;1;1;1
+141;1;1;1
+142;1;1;1
+143;1;1;1
+144;1;1;1
+145;1;1;1
+146;1;0;1
+147;1;0;1
+148;1;0;1
+149;1;0;1
+150;1;0;1
+151;1;0;1
+152;1;0;1
+153;1;0;1
+154;1;0;1
+155;1;1;1
+156;1;1;1
+157;1;1;1
+158;1;1;1
+159;1;1;1
+160;1;1;1
+161;1;1;1
+162;1;1;1
+163;1;1;1
+164;1;1;1
+165;1;1;1
+166;1;1;1
+167;1;1;1
+168;0;0;1
+169;0;0;1
+170;0;0;1
+171;0;0;1
+172;0;0;1
+173;0;0;1
+174;0;0;1
+175;0;0;1
+176;0;0;1
+177;0;0;1
+178;0;0;1
+179;0;0;1
+180;0;0;1
+181;0;0;1
+182;0;0;1
+183;0;0;1
+184;0;0;1
+185;0;0;1
+186;0;0;1
+187;0;0;1
+188;0;0;1
+;0;0;
+;0;0;
+;0;0;
+;0;0;
